requirements.txt
fastapi
google.generativeai
supabase
uvicorn
python-dotenv
